moved more files around to the correct location
